+++ Source/Alternet.UI.Common/Drawing/Skia/SkiaUtils.cs
@@ -14,11 +14,31 @@ namespace Alternet.Drawing
     /// </summary>
     public static class SkiaUtils
     {
+        internal static readonly EnumArray<InterpolationMode, SKPaint> InterpolationModePaints = new();
+
         private static string[]? fontFamilies;
         private static FontSize defaultFontSize = 12;
         private static string? defaultMonoFontName;
         private static SKFont? defaultSkiaFont;
         private static SKTypeface? defaultTypeFace;
+
+        static SkiaUtils()
+        {
+            InterpolationModePaints[InterpolationMode.None] = CreatePaints(SKFilterQuality.None);
+            InterpolationModePaints[InterpolationMode.LowQuality] = CreatePaints(SKFilterQuality.Low);
+            InterpolationModePaints[InterpolationMode.HighQuality] = CreatePaints(SKFilterQuality.High);
+            InterpolationModePaints[InterpolationMode.MediumQuality] = CreatePaints(SKFilterQuality.Medium);
+
+            SKPaint CreatePaints(SKFilterQuality quality)
+            {
+                SKPaint result = new()
+                {
+                    FilterQuality = quality,
+                };
+
+                return result;
+            }
+        }
 
         /// <summary>
         /// Gets or sets default font for use with SkiaSharp.
@@ -186,6 +206,28 @@ namespace Alternet.Drawing
         public static SKFont CreateDefaultFont()
         {
             return new SKFont(DefaultTypeFace, (float)DefaultFontSize);
+        }
+
+        /// <summary>
+        /// Converts <see cref="InterpolationMode"/> to <see cref="SKFilterQuality"/>.
+        /// </summary>
+        /// <param name="interpolationMode">Value to convert.</param>
+        /// <returns></returns>
+        public static SKFilterQuality ToSkia(this InterpolationMode interpolationMode)
+        {
+            switch (interpolationMode)
+            {
+                case InterpolationMode.None:
+                    return SKFilterQuality.None;
+                case InterpolationMode.LowQuality:
+                    return SKFilterQuality.Low;
+                case InterpolationMode.MediumQuality:
+                    return SKFilterQuality.Medium;
+                case InterpolationMode.HighQuality:
+                    return SKFilterQuality.High;
+                default:
+                    return SKFilterQuality.None;
+            }
         }
 
         /// <summary>
@@ -405,27 +447,6 @@ namespace Alternet.Drawing
             canvas.Scale(dpi / 96.0f);
             onRender(surface);
             canvas.Flush();
-        }
-
-        internal static void SampleDrawText(SKCanvas canvas, string text, SKRect rect, SKPaint paint)
-        {
-            float spaceWidth = paint.MeasureText(" ");
-            float wordX = rect.Left;
-            float wordY = rect.Top + paint.TextSize;
-            foreach (string word in text.Split(' '))
-            {
-                float wordWidth = paint.MeasureText(word);
-                if (wordWidth <= rect.Right - wordX)
-                {
-                    canvas.DrawText(word, wordX, wordY, paint);
-                    wordX += wordWidth + spaceWidth;
-                }
-                else
-                {
-                    wordY += paint.FontSpacing;
-                    wordX = rect.Left;
-                }
-            }
         }
     }
 }